requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
Deprecated==1.2.14
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
limits==3.13.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.12.1
